Move pr backend tests to separate phi tests
