+++ tests/CMakeLists.txt
@@ -3,6 +3,7 @@ cmake_minimum_required(VERSION 3.11)
 add_subdirectory(cc)
 add_subdirectory(rf)
 add_subdirectory(td)
+add_subdirectory(phi)
 add_subdirectory(pr)
 add_subdirectory(rlog)
 add_subdirectory(nx)
