+++ tests/phi/main.cc
@@ -1,0 +1,3 @@
+#include <nexus/run.hh>
+
+int main(int argc, char** argv) { return nx::run(argc, argv); }
